requirements.txt
streamlit
openai
spotipy
python-dotenv
textblob
nltk
